requirements.txt
attrs==24.2.0
black==24.8.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
distlib==0.3.8
dnspython==2.6.1
filelock==3.15.4
Flask==3.0.3
Flask-RESTful==0.3.10
greenlet==3.0.3
identify==2.6.0
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
mypy-extensions==1.0.0
nested-lookup==0.2.25
nodeenv==1.9.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
packaging==24.1
pathable==0.4.3
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
pymongo==4.8.0
pyproject-api==1.7.1
pytest==8.3.2
pytest-cov==5.0.0
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rpds-py==0.20.0
six==1.16.0
sourcery-cli==1.18.0
SQLAlchemy==2.0.32
tox==4.18.0
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
Werkzeug==3.0.4
